Created Fir map
